+++ src/features/navigation/RootNavigator/Header/NavItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components/native'
 
-import { menu } from 'features/navigation/TabBar/routes'
+import { menu } from 'features/navigation/TabBar/menu'
 import { TabRouteName } from 'features/navigation/TabBar/types'
 import { InternalTouchableLink } from 'ui/components/touchableLink/InternalTouchableLink'
 import { InternalNavigationProps } from 'ui/components/touchableLink/types'
