+++ src/features/navigation/TabBar/routes.ts
@@ -1,5 +1,4 @@
 import { LinkingOptions } from '@react-navigation/native'
-import { Platform } from 'react-native'
 
 import { Bookings } from 'features/bookings/pages/Bookings/Bookings'
 import { withAsyncErrorBoundary } from 'features/errors/hocs/withAsyncErrorBoundary'
@@ -63,15 +62,6 @@ const tabRouteNames = routes.map((route) => route.name)
 export function isTabScreen(screen: ScreenNames): screen is TabRouteName {
   // @ts-expect-error : ScreenNames is not necessarily a screen in tabRouteNames
   return tabRouteNames.includes(screen)
-}
-
-const isWeb = Platform.OS === 'web'
-export const menu: Record<TabRouteName, { displayName: string; accessibilityLabel?: string }> = {
-  Home: { displayName: 'Accueil', accessibilityLabel: 'Accueil' },
-  Search: { displayName: 'Recherche', accessibilityLabel: 'Rechercher des offres' },
-  Bookings: { displayName: 'Réservations', accessibilityLabel: 'Mes réservations' },
-  Favorites: { displayName: 'Favoris', accessibilityLabel: isWeb ? undefined : 'Mes favoris' },
-  Profile: { displayName: 'Profil', accessibilityLabel: 'Mon profil' },
 }
 
 const { screensConfig: tabScreensConfig, Screens: TabScreens } = getScreensAndConfig(
